move significant_digit
